# docs/reference/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kingshot Alliance API",
    "version": "1.0.0",
    "description": "Bounded Alliance reads plus verified, idempotent external-actor participation adapters. Business writes remain owner-context actions."
  },
  "servers": [{ "url": "/" }],
  "security": [{ "bearerToken": [] }],
  "paths": {
...
    "/api/v1/actor-links/claims": {
      "post": {
        "operationId": "claimExternalActorLink",
        "x-required-scope": "actor-links:write",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["provider", "external_subject", "code"],
                "properties": {
                  "provider": { "$ref": "#/components/schemas/ExternalActorProvider" },
                  "external_subject": { "type": "string", "pattern": "^[1-9][0-9]{4,24}$" },
                  "code": { "type": "string", "maxLength": 20 }
                }
              }
            }
          }
        },
        "responses": { "200": { "$ref": "#/components/responses/ActionResponse" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ks_live_<prefix>.<secret>"
      }
    },
...
    "schemas": {
      "ExternalActorProvider": { "type": "string", "enum": ["discord", "telegram"] },
...
    },
    "responses": {
...
      "ActionResponse": {
        "description": "Successful actor-scoped action or safe replay.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["data"],
              "properties": {
                "data": { "type": "object" },
                "meta": {
                  "type": "object",
                  "properties": { "replayed": { "type": "boolean" } }
                }
              }
            }
          }
        }
      }
    }
  }
}